requirements.txt
fastapi
uvicorn[standard]
Pillow
pytesseract
pdf2image

